+++ docs/gatsby-config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  siteMetadata: {
+    title: 'react-responsive-modal',
+    author: 'Léo Pradel',
+  },
+  plugins: [
+    'gatsby-plugin-react-helmet',
+  ],
+};
