+++ docs/package.json
@@ -7,6 +7,11 @@
   },
   "dependencies": {
     "gatsby": "^1.0.0",
-    "gatsby-link": "^1.0.0"
+    "gatsby-link": "^1.0.0",
+    "gatsby-plugin-react-helmet": "^1.0.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-helmet": "^5.1.3"
   }
 }
