+++ package.json
@@ -7,7 +7,7 @@
     "build": "babel src --out-dir lib",
     "build:watch": "npm run build -- --watch",
     "build:examples": "browserify examples/app.js -o examples/app.bundle.js -t [ babelify --presets [ es2015 react ] ]",
-    "lint": "eslint src",
+    "lint": "eslint src docs/src",
     "test": "npm run lint && npm run build",
     "format": "prettier --single-quote --trailing-comma es5 --write \"src/**/*.js\" \"docs/src/**/*.{js,css}\""
   },
@@ -38,6 +38,7 @@
   "devDependencies": {
     "babel-cli": "^6.24.1",
     "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babelify": "^7.3.0",
